pyproject.toml
[project]
name = "auction-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.9.1",
    "django>=5.2.4",
    "djangorestframework==3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=11.3.0",
    "ruff>=0.12.7",
    "sqlparse==0.5.3",
]
